requirements.txt
flask
pymongo
opencv-python
numpy
pillow
imagehash
cryptography
boto3
